Fix redundant cast in VirtualReel get() method
Removed an unnecessary cast to int in the get() method, as the data array already contains integers. This improves code readability and ensures cleaner implementation without altering functionality.

+++ src/main/java/wtd/slotsengine/slots/machines/reels/VirtualReel.java
@@ -22,7 +22,7 @@ public class VirtualReel implements IReel {
     }
 
     public int get(int position) {
-        return (int) data[position % data.length];
+        return data[position % data.length];
     }
 
     public int size() {
